+++ WorkShop4/ProductFrm.cs
@@ -67,9 +67,19 @@ namespace WorkShop4
 
         private void RemoveBtn_Click(object sender, EventArgs e)
         {
-            context.Products.Remove(currentProduct);
-            context.SaveChanges();
-            Display();
+            DialogResult button =
+                MessageBox.Show(
+                    "Are you sure that you want to remove the selected record?",
+                    "Remove Item",
+                    MessageBoxButtons.OKCancel,
+                    MessageBoxIcon.Question,
+                    MessageBoxDefaultButton.Button2);
+            if (button == DialogResult.OK)
+            {
+                context.Products.Remove(currentProduct);
+                context.SaveChanges();
+                Display();
+            }
         }
 
         private void ExitBtn_Click(object sender, EventArgs e)
