+++ WorkShop4/ProductSupplierFrm.cs
@@ -61,9 +61,19 @@ namespace WorkShop4
 
         private void removeSupplierBtn_Click(object sender, EventArgs e)
         {
-            context.ProductsSuppliers.Remove(currentProductSupplier);
-            context.SaveChanges();
-            Display();
+            DialogResult button =
+                MessageBox.Show(
+                    "Are you sure that you want to remove the selected record?",
+                    "Remove Item",
+                    MessageBoxButtons.OKCancel,
+                    MessageBoxIcon.Question,
+                    MessageBoxDefaultButton.Button2);
+            if (button == DialogResult.OK)
+            {
+                context.ProductsSuppliers.Remove(currentProductSupplier);
+                context.SaveChanges();
+                Display();
+            }
         }
 
         private void exitSupplierBtn_Click(object sender, EventArgs e)
